+++ public/class-gforms-bootstrapper-public.php
@@ -118,4 +118,13 @@ class Gforms_Bootstrapper_Public {
 		return '<input type="submit" class="gform_button button btn btn-primary" id="gform_submit_button_' . $form['id'] . '" value="' . $form['button']['text'] . '">';
 	}
 
+	/**
+	 * Customize the error validation message
+	 *
+	 * @since    2.0.2
+	 */
+	public function custom_validation_message( $message, $form ) {
+		return '<div class="alert alert-warning alert-dismissible fade show">' . $message . '<button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>';
+	}
+
 }